# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for OnePenny server.",
        "title": "OnePenny API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/applications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页获取当前登录用户提交的所有赏金申请",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "列出当前用户的赏金申请",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页大小",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "申请列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/application.ApplicationResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "登录用户向指定赏金任务提交申请",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "提交赏金申请",
                "parameters": [
                    {
                        "description": "申请信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.SubmitApplicationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "申请提交成功",
                        "schema": {
                            "$ref": "#/definitions/application.ApplicationResponse"
                        }
                    },
                    "400": {
                        "description": "参数格式错误",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/applications/{application_id}/approve": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "发布方批准某个申请，并说明理由，悬赏令进入进行中状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "批准申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "申请 ID",
                        "name": "application_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "批准理由",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.decisionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.ApplicationDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/applications/{application_id}/reject": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "发布方拒绝某个申请，并说明理由，悬赏令保持 Created 状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "拒绝申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "申请 ID",
                        "name": "application_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "拒绝理由",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.decisionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.ApplicationDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/applications/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据申请 ID 查询单个申请信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "获取赏金申请详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "申请 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/application.ApplicationResponse"
                        }
                    },
                    "400": {
                        "description": "无效的申请 ID",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "申请不存在",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据申请 ID 删除当前用户的赏金申请",
                "tags": [
                    "application"
                ],
                "summary": "删除赏金申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "申请 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "无效的申请 ID",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/application.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "application.ApplicationResponse": {
            "type": "object",
            "properties": {
                "attachment_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bounty_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "proposal": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "application.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "application.SubmitApplicationRequest": {
            "type": "object",
            "required": [
                "bounty_id",
                "proposal"
            ],
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bounty_id": {
                    "type": "string"
                },
                "proposal": {
                    "type": "string"
                }
            }
        },
        "application.decisionRequest": {
            "type": "object",
            "required": [
                "reason"
            ],
            "properties": {
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "service.ApplicationDTO": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bounty_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "proposal": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    }
}